+++ src/org/pipservices3/commons/convert/LongConverter.java
@@ -87,7 +87,7 @@ public class LongConverter {
      */
     public static long toLongWithDefault(Object value, long defaultValue) {
         Long result = toNullableLong(value);
-        return result != null ? (long) result : defaultValue;
+        return result != null ? result : defaultValue;
     }
 
 }
